Fix compiler warnings

+++ utility/SdFile.cpp
@@ -896,7 +896,7 @@ uint8_t SdFile::rmRfStar(void) {
       if (!f.remove()) return false;
     }
     // position to next entry if required
-    if (curPosition_ != (32*(index + 1))) {
+    if (curPosition_ != (uint32_t)(32*(index + 1))) {
       if (!seekSet(32*(index + 1))) return false;
     }
   }
